Generate protos (#4)
This is a first attempt at setting up the build files and generating the protos.

Since the generated protos are part of the repo, the generation itself is not part of the build pipeline and is set up as an standalone project in the `generateProtos` directory. Regenerating them can be done manually as explained in the README.

While the issue mentions a makefile for proto generation, I took inspiration from [opentelemetry-proto-java](https://github.com/open-telemetry/opentelemetry-proto-java) and used a gradle script instead. I can change that if needed.

Note also that the generated code is not published as a library. I suggest doing this in a separate PR.

Resolves #1